Advertise AirPlay receiver on the configured stream interface
The AirPlay receiver's shairport-sync instance advertised its mDNS service on all interfaces, so on multi-homed or Docker host-network setups it could announce an unreachable address (e.g. a Docker bridge). Pin the advertisement to the stream server's bind interface when one is configured, mirroring the Spotify Connect provider. The shared interface_name_for_ip helper is promoted to helpers/util.py.
