+++ music_assistant/helpers/util.py
@@ -937,6 +937,24 @@ async def get_ip_addresses(include_ipv6: bool = False) -> tuple[str, ...]:
     return await asyncio.to_thread(call)
 
 
+def interface_name_for_ip(ip: str) -> str | None:
+    """
+    Return the name of the network interface that holds the given IP, or None.
+
+    Used to map a bind/publish IP to its interface name for components that select
+    their mDNS/zeroconf advertisement interface by name (e.g. shairport-sync and
+    go-librespot), so the advertisement stays on the intended network.
+
+    :param ip: The IPv4/IPv6 address to look up.
+    """
+    for adapter in ifaddr.get_adapters():
+        for ip_config in adapter.ips:
+            addr = ip_config.ip if isinstance(ip_config.ip, str) else ip_config.ip[0]
+            if addr == ip:
+                return adapter.name
+    return None
+
+
 async def get_primary_ip_address() -> str | None:
     """Return the primary IP address of the system."""
