+++ music_assistant/providers/airplay_receiver/bin/shairport-sync.conf
@@ -8,8 +8,10 @@ general =
 	output_backend = "pipe";
 	default_airplay_volume = {DEFAULT_VOLUME};
 	port = {PORT};
-	// Using tinysvcmdns (embedded mDNS, no external daemon required)
-};
+	// Using tinysvcmdns (embedded mDNS, no external daemon required).
+	// The line below is filled in with an `interface = "..."` directive when the
+	// stream server is bound to a specific interface, else it is left empty.
+{INTERFACE_LINE}};
 
 metadata =
 {
